refactor: extend common vitest config between test projects
